Shared helper methods splitted.
